+++ src/ops/matmul/cuda/matmul_cuda.h
@@ -5,6 +5,7 @@
 
 typedef struct MatmulCudaDescriptor {
     Device device;
+    MatmulCudaDescriptor(Device device);
 } MatmulCudaDescriptor;
 
 void matmul_nv_gpu_f16(MutTensor c, float beta, ConstTensor a, ConstTensor b, float alpha, void *stream);
